+++ src/AttackGraphNodeShape.ts
@@ -42,7 +42,7 @@ export class AttackGraphNodeShape extends mxRectangleShape {
     }
 
     const graph = this.state?.view.graph;
-    const globalAttributes = AttributeRenderer.rootAttributes(graph as unknown as Draw.EditorGraph).getGlobalAttributes();
+    const globalAttributes = AttributeRenderer.rootAttributes().getGlobalAttributes();
     let tempAttributes = Object.entries(attributes);
     if (globalAttributes !== null) {
       const renderableAttributes = RootAttributeProvider.getRenderableAttributes(globalAttributes);
@@ -70,7 +70,7 @@ export class AttackGraphNodeShape extends mxRectangleShape {
 
       let icon = `data:image/svg+xml;utf8,${Framework7Icons.Icons.question}`;
       if (graph !== undefined) {
-        const globalAttribute = AttributeRenderer.rootAttributes(graph as unknown as Draw.EditorGraph).getGlobalAttribute(key);
+        const globalAttribute = AttributeRenderer.rootAttributes().getGlobalAttribute(key);
         if (globalAttribute !== null) {
           if (Framework7Icons.Icons[globalAttribute.iconName] !== undefined) {
             icon = `data:image/svg+xml;utf8,${Framework7Icons.Icons[globalAttribute.iconName]}`;
@@ -134,18 +134,21 @@ export class AttackGraphNodeShape extends mxRectangleShape {
   private drawLabelShape(label: string, c: import('mxgraph').mxAbstractCanvas2D, w: number, fillColor: string, fontColor: string) {
     if (this.state) {
       const opacity = mxUtils.getValue(this.state.style, 'opacity', '100') as number;
-      const bubbleDiameter = 20;
+      const squareDiameter = 20;
       c.setFillColor(fillColor);
-      c.setStrokeColor('#000');
-      c.ellipse(w - (bubbleDiameter * 0.5), - (bubbleDiameter * 0.5), bubbleDiameter, bubbleDiameter);
+      c.setStrokeColor('#000000');
+      c.rect(w - squareDiameter, 0, squareDiameter, squareDiameter);
       c.fillAndStroke();
 
       c.setAlpha(opacity / 100);
       c.setFontStyle(mxConstants.DEFAULT_FONTSTYLE.toString());
       c.setFontColor(fontColor);
       c.setFontSize(13);
-
-      c.text(w, 0, 0, 0, label === 'NaN' ? '!' : label, mxConstants.ALIGN_CENTER, mxConstants.ALIGN_MIDDLE, 0, null, 0, 0, 0);
+      c.text(
+        w - squareDiameter * 0.5,
+        squareDiameter * 0.5,
+        0, 0, label === 'NaN' ? '!' : label, mxConstants.ALIGN_CENTER, mxConstants.ALIGN_MIDDLE, 0, null, 0, 0, 0
+      );
     }
   }