+++ src/AttributeRenderer.ts
@@ -7,6 +7,7 @@ import { AttackGraphSettings } from './AttackGraphSettings';
 import { GraphUtils } from './GraphUtils';
 import { AttackgraphFunction, CellDataCollection, ChildCellData, ChildCellDataCollection, GlobalAttribute, GlobalAttributeDict, KeyValuePairs } from './Model';
 import { Menubar } from './Menubar';
+import { CellStyles } from './Analysis/CellStyles';
 
 export class AttributeRenderer {
   private static _sensitivityAnalysisEnabled = false;
@@ -27,11 +28,11 @@ export class AttributeRenderer {
     }
   }
 
-  static rootAttributes(graph: Draw.EditorGraph): RootAttributeProvider {
+  static rootAttributes(): RootAttributeProvider {
     if (this._sensitivityAnalysisEnabled) {
-      return new SensitivityAnalysisRootAttributeProvider(graph);
+      return new SensitivityAnalysisRootAttributeProvider();
     } else {
-      return new RootAttributeProvider(graph);
+      return new RootAttributeProvider();
     }
   }
 
@@ -51,7 +52,7 @@ export class AttributeRenderer {
     if (AttackGraphSettings.isAttackGraph(ui.editor.graph)) {
       if (GraphUtils.isTree(cell)) {
         Menubar.increaseUnfinishedWorkers(ui);
-        await this.updateCellValuesUpwards(this.nodeAttributes(cell), AttributeRenderer.rootAttributes(ui.editor.graph), worker);
+        await this.updateCellValuesUpwards(this.nodeAttributes(cell), AttributeRenderer.rootAttributes(), worker, ui);
         Menubar.decreaseUnfinishedWorkers(ui);
       } else {
         mxUtils.alert('Cannot recalculate the graph as it contains loops!');
@@ -59,31 +60,49 @@ export class AttributeRenderer {
     }
   }
 
-  private static async updateCellValuesUpwards(cell: NodeAttributeProvider, graph: RootAttributeProvider, worker: AsyncWorker): Promise<void> {
-    const globalDefaultAttributes = graph.getGlobalAttributes();
+  private static async updateCellValuesUpwards(cell: NodeAttributeProvider, root: RootAttributeProvider, worker: AsyncWorker, ui: Draw.UI): Promise<void> {
+    const globalDefaultAttributes = root.getGlobalAttributes();
     const globalDefaultAttributesDict = this.transformGlobalAttributesToGlobalAttributesDict(globalDefaultAttributes || []) as GlobalAttributeDict;
+    const incomingEdges = cell.cell.edges?.filter(x => x.target === cell.cell && x.source) || [];
 
     const childValues = this.getChildValues(cell);
-    const aggregationFunction = cell.resolveAggregationFunction(graph);
+    const aggregationFunction = cell.resolveAggregationFunction(root);
 
     const localAttributes = cell.getCellValues();
+    let aggregatedValues: KeyValuePairs = {};
+    let aggregate = true;
 
-    const aggregatedValues = await this.aggregateAttributes({
-      globalAttributes: globalDefaultAttributesDict,
-      childAttributes: childValues,
-      localAttributes: localAttributes,
-      id: cell.getCellId()
-    }, aggregationFunction, worker);
+    // Link to another page
+    if (new CellStyles(cell.cell).isLinkNode()) {
+      if (cell.isLeave()) { // source link node
+        const refCell = cell.getReferencedCell();        
+        if (refCell) {
+          aggregatedValues = refCell.getAggregatedCellValues();
+          aggregate = false;
+        } else {
+          aggregatedValues['_error'] = 'true';
+        }
+      }
+    }
+
+    if (aggregate) {
+      const values = await this.aggregateAttributes({
+        globalAttributes: globalDefaultAttributesDict,
+        childAttributes: childValues,
+        localAttributes: localAttributes,
+        id: cell.getCellId()
+      }, aggregationFunction, worker);
+      aggregatedValues = { ...values, ...aggregatedValues }; // Include possible link errors
+    }
     cell.setAggregatedCellValues(aggregatedValues);
 
-    const labelFunction = cell.resolveComputedAttributesFunction(graph);
+    const labelFunction = cell.resolveComputedAttributesFunction(root);
 
     const cellAttributes = cell.getCellValues();
     const computedAttributes = await this.recalculateCellLabel({ globalAttributes: globalDefaultAttributesDict, cellAttributes: { ...cellAttributes, ...aggregatedValues } }, labelFunction, worker);
     cell.setComputedAttributesForCell(computedAttributes);
 
-    const incomingEdges = cell.cell.edges?.filter(x => x.target === cell.cell && x.source) || [];
-    await Promise.all(incomingEdges.map(x => this.updateCellValuesUpwards(this.nodeAttributes(x.source), graph, worker)));
+    await Promise.all(incomingEdges.map(x => this.updateCellValuesUpwards(this.nodeAttributes(x.source), root, worker, ui)));
   }
 
   public static transformGlobalAttributesToGlobalAttributesDict(attributes: GlobalAttribute[]): { [name: string]: KeyValuePairs } {