+++ types/drawio/index.d.ts
@@ -5,10 +5,15 @@ declare namespace Draw {
 
   class UI {
     actions: Actions;
+    currentPage: DiagramPage;
+    pages: DiagramPage[];
     menubar: Menubar;
     menus: UIMenus;
     copiedValue: null;
     getCurrentFile(): File;
+    getPageIndex(page: DiagramPage): number;
+    getPageById(id: string, pages?: DiagramPage[]): DiagramPage | null;
+    selectPage(page: DiagramPage, quiet?: boolean, viewState?: Object): void;
     sidebar: Sidebar;
     footerContainer: HTMLElement;
     fileNode: FileNode;
@@ -17,6 +22,14 @@ declare namespace Draw {
     showDialog(elt, w, h, modal, closable, onClose?, noScroll?, transparent?, onResize?, ignoreBgClick?);
     confirm(msg: string, okFn: () => void, cancelFn?: () => void, okLabel?: string, cancelLabel?: string, closable?: boolean): void;
     hideDialog();
+  }
+
+  class DiagramPage {
+    root: import('mxgraph').mxCell;
+    viewState: Object;
+    needsUpdate: boolean;
+    getName(): string;
+    getId(): string;
   }
 
   class Actions {
@@ -110,6 +123,13 @@ declare namespace Draw {
   class File {
     getTitle();
   }
+}
+
+declare class MovePage {
+  ui: UI;
+  oldIndex: number;
+  newIndex: number;
+  execute(): void;
 }
 
 declare class Graph {